Moved PAN, PANFile and exceptions to separate classes
