+++ exceptions.py
@@ -1,0 +1,7 @@
+
+class MSGException(BaseException):
+    pass
+
+
+class PSTException(BaseException):
+    pass
